+++ src/filters/boolean_filter.ts
@@ -24,6 +24,7 @@ export enum BooleanFilterName {
     // Server Side filters work a bit differently. The API doesn't return toots that match the filter
     // for authenticated requests but for unauthenticated requests (e.g. pulling trending toots from
     // other servers) it does so we have to manually filter them out.
+    // TODO: we now filter these out in isValidForFeed() so this stuff is probably not needed
     SERVER_SIDE_FILTERS = StorageKey.SERVER_SIDE_FILTERS,
 };