+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Filter out toots matching the server side filters altogether (in particular from the list of trending toots)
 * Use split() method to separate `Toot`s needing completion from those that are complete
 
 ### v0.41.3